Fix includes and config types in quantum optimizer
